# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GlucoJack - Type 1 Diabetes insulin tracking and decision-support API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type EntryMealType = (typeof EntryMealType)[keyof typeof EntryMealType];

export const EntryMealType = {
  FAST_CARBS: "FAST_CARBS",
  HIGH_FAT: "HIGH_FAT",
  HIGH_PROTEIN: "HIGH_PROTEIN",
  BALANCED: "BALANCED",
} as const;

/**
 * @nullable
 */
export type EntryEvaluation =
  | (typeof EntryEvaluation)[keyof typeof EntryEvaluation]
  | null;

export const EntryEvaluation = {
  GOOD: "GOOD",
  OVERDOSE: "OVERDOSE",
  UNDERDOSE: "UNDERDOSE",
  CHECK_CONTEXT: "CHECK_CONTEXT",
} as const;

export interface Entry {
  id: number;
  timestamp: string;
  /** Blood glucose before meal (mg/dL) */
  glucoseBefore: number;
  /**
   * Blood glucose after meal (mg/dL)
   * @nullable
   */
  glucoseAfter?: number | null;
  /** Carbohydrate intake in grams */
  carbsGrams: number;
  /** Bolus insulin units administered */
  insulinUnits: number;
  mealType: EntryMealType;
  /** @nullable */
  mealDescription?: string | null;
  /**
   * Minutes between before/after readings
   * @nullable
   */
  timeDifferenceMinutes?: number | null;
  /**
   * Glucose change (after - before)
   * @nullable
   */
  delta?: number | null;
  /**
   * Rate of change (mg/dL per hour)
   * @nullable
   */
  speed?: number | null;
  /** @nullable */
  evaluation?: EntryEvaluation;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface GetEntriesResponse {
  entries: Entry[];
  total: number;
}

export type CreateEntryBodyMealType =
  (typeof CreateEntryBodyMealType)[keyof typeof CreateEntryBodyMealType];

export const CreateEntryBodyMealType = {
  FAST_CARBS: "FAST_CARBS",
  HIGH_FAT: "HIGH_FAT",
  HIGH_PROTEIN: "HIGH_PROTEIN",
  BALANCED: "BALANCED",
} as const;

export interface CreateEntryBody {
  timestamp?: string;
  glucoseBefore: number;
  /** @nullable */
  glucoseAfter?: number | null;
  carbsGrams: number;
  insulinUnits: number;
  mealType: CreateEntryBodyMealType;
  /** @nullable */
  mealDescription?: string | null;
  /** @nullable */
  timeDifferenceMinutes?: number | null;
  /** @nullable */
  notes?: string | null;
}

export interface ImportBatchBody {
  entries: CreateEntryBody[];
}

export interface ImportBatchResponse {
  imported: number;
  entries: Entry[];
}

export interface EvaluationBreakdown {
  GOOD: number;
  OVERDOSE: number;
  UNDERDOSE: number;
  CHECK_CONTEXT: number;
}

export interface DashboardStats {
  /** Overall control score 0-100 */
  controlScore: number;
  /** Percentage of entries with hypoglycemia risk */
  hypoRate: number;
  /** Percentage of entries with hyperglycemia spike */
  spikeRate: number;
  totalEntries: number;
  /** Percentage of GOOD evaluations */
  goodRate: number;
  /** @nullable */
  avgGlucoseBefore?: number | null;
  /** @nullable */
  avgGlucoseAfter?: number | null;
  /** @nullable */
  avgCarbsGrams?: number | null;
  /** @nullable */
  avgInsulinUnits?: number | null;
  recentEntries: Entry[];
  evaluationBreakdown: EvaluationBreakdown;
}

export type MealPatternMealType =
  (typeof MealPatternMealType)[keyof typeof MealPatternMealType];

export const MealPatternMealType = {
  FAST_CARBS: "FAST_CARBS",
  HIGH_FAT: "HIGH_FAT",
  HIGH_PROTEIN: "HIGH_PROTEIN",
  BALANCED: "BALANCED",
} as const;

export interface MealPattern {
  mealType: MealPatternMealType;
  count: number;
  avgCarbsGrams: number;
  avgInsulinUnits: number;
  /** @nullable */
  avgDelta?: number | null;
  /** @nullable */
  avgSpeed?: number | null;
  goodRate: number;
  /** Average insulin units per 10g carbs */
  insulinToCarb: number;
}

export interface GetMealPatternsResponse {
  patterns: MealPattern[];
}

/**
 * @nullable
 */
export type GlucosePointEvaluation =
  | (typeof GlucosePointEvaluation)[keyof typeof GlucosePointEvaluation]
  | null;

export const GlucosePointEvaluation = {
  GOOD: "GOOD",
  OVERDOSE: "OVERDOSE",
  UNDERDOSE: "UNDERDOSE",
  CHECK_CONTEXT: "CHECK_CONTEXT",
} as const;

export interface GlucosePoint {
  timestamp: string;
  glucoseBefore: number;
  /** @nullable */
  glucoseAfter?: number | null;
  /** @nullable */
  evaluation?: GlucosePointEvaluation;
}

export interface GetGlucoseTrendResponse {
  points: GlucosePoint[];
}

export type RecommendationRequestMealType =
  (typeof RecommendationRequestMealType)[keyof typeof RecommendationRequestMealType];

export const RecommendationRequestMealType = {
  FAST_CARBS: "FAST_CARBS",
  HIGH_FAT: "HIGH_FAT",
  HIGH_PROTEIN: "HIGH_PROTEIN",
  BALANCED: "BALANCED",
} as const;

export interface RecommendationRequest {
  carbsGrams: number;
  glucoseBefore: number;
  mealType: RecommendationRequestMealType;
}

/**
 * Based on how many similar entries exist
 */
export type RecommendationConfidence =
  (typeof RecommendationConfidence)[keyof typeof RecommendationConfidence];

export const RecommendationConfidence = {
  HIGH: "HIGH",
  MEDIUM: "MEDIUM",
  LOW: "LOW",
} as const;

export type RecommendationTiming =
  (typeof RecommendationTiming)[keyof typeof RecommendationTiming];

export const RecommendationTiming = {
  BEFORE_MEAL: "BEFORE_MEAL",
  SPLIT_DOSE: "SPLIT_DOSE",
  WITH_MEAL: "WITH_MEAL",
} as const;

export interface Recommendation {
  /** Suggested bolus insulin units */
  recommendedUnits: number;
  minUnits: number;
  maxUnits: number;
  /** Based on how many similar entries exist */
  confidence: RecommendationConfidence;
  timing: RecommendationTiming;
  /** Human-readable explanation */
  reasoning: string;
  /** Number of similar past entries used */
  basedOnEntries: number;
}

export type GetEntriesParams = {
  limit?: number;
  offset?: number;
  mealType?: GetEntriesMealType;
};

export type GetEntriesMealType =
  (typeof GetEntriesMealType)[keyof typeof GetEntriesMealType];

export const GetEntriesMealType = {
  FAST_CARBS: "FAST_CARBS",
  HIGH_FAT: "HIGH_FAT",
  HIGH_PROTEIN: "HIGH_PROTEIN",
  BALANCED: "BALANCED",
} as const;

export type GetGlucoseTrendParams = {
  days?: number;
};